requirements.txt
flask
anthropic
openai
google-generativeai
pyswisseph
geopy
timezonefinder
pytz
